refactor(plugins): move healthPlugin to src/plugins/health/ (CP-1)
First plugin migrated under the new convention.

- src/lib/health/index.ts        -> src/plugins/health/health-plugin.ts
- src/lib/health/health.model.ts -> src/plugins/health/health.model.ts
- src/lib/health/__tests__/health.test.ts
    -> src/plugins/health/__tests__/health-plugin.test.ts
- 3 import sites updated: src/index.ts, src/test/support/app.ts,
  src/test/helpers/app.ts

No behavior change. 7 health tests pass in the new location.
