+++ src/index.ts
@@ -11,7 +11,6 @@ import { betterAuthPlugin, OpenAPI } from "./lib/auth-plugin";
 import { parseOrigins } from "./lib/cors";
 import { cronPlugin } from "./lib/cron-plugin";
 import { errorPlugin } from "./lib/errors/error-plugin";
-import { healthPlugin } from "./lib/health";
 import { logger, loggerPlugin } from "./lib/logger";
 import { setupGracefulShutdown } from "./lib/shutdown/shutdown";
 import { adminController } from "./modules/admin";
@@ -23,6 +22,7 @@ import { organizationController } from "./modules/organizations";
 import { paymentsController } from "./modules/payments";
 import { registerPaymentListeners } from "./modules/payments/hooks/listeners";
 import { publicController } from "./modules/public";
+import { healthPlugin } from "./plugins/health/health-plugin";
 
 const corsOrigins = parseOrigins(env.CORS_ORIGIN);
